frontend/fix: #5574 Moved MobilityCallAPI to common and decoupled mobility-common from mobility-app
