+++ Frontend/android-native/mobility-app/build.gradle
@@ -43,6 +43,7 @@ android {
 def cameraxVersion = "1.1.0-alpha05"
 
 dependencies {
+    implementation project(':mobility-common')
     implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation 'com.google.android.material:material:1.6.0'
     implementation 'com.google.android.play:app-update:2.0.0'
