+++ Frontend/android-native/mobility-app/src/main/java/in/juspay/mobility/app/Utils.java
@@ -36,94 +36,6 @@ public class Utils {
 
     private static final String UTILS = "UTILS";
 
-    private static final ArrayList<CallBack> callBack = new ArrayList<>();
-
-    public static void registerCallback(CallBack notificationCallback) {
-        callBack.add(notificationCallback);
-    }
-
-    public static void deRegisterCallback(CallBack notificationCallback) {
-        callBack.remove(notificationCallback);
-    }
-
-    public static void captureImage(@Nullable Intent data, Activity activity, Context context) {
-        try {
-            Uri imageUri;
-            SharedPreferences sharedPref = context.getSharedPreferences(context.getString(R.string.preference_file_key), Context.MODE_PRIVATE);
-            if (data == null || data.getData() == null) { //Camera
-                File image = new File(context.getFilesDir(), "IMG_" + sharedPref.getString(context.getResources().getString(R.string.TIME_STAMP_FILE_UPLOAD), "null") + ".jpg");
-                imageUri = FileProvider.getUriForFile(context, context.getPackageName() + ".provider", image);
-            } else { // storage
-                imageUri = data.getData();
-            }
-            startCropImageActivity(imageUri, activity);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
-    public static void startCropImageActivity(Uri imageUri, Activity activity) {
-        CropImage.activity(imageUri)
-                .setAllowFlipping(false)
-                .start(activity);
-    }
-
-    public static void encodeImageToBase64(@Nullable Intent data, Context context, @Nullable Uri imageData) {
-        FirebaseAnalytics mFirebaseAnalytics = FirebaseAnalytics.getInstance(context);
-        try {
-            Uri fileUri;
-            if(imageData == null) {
-                CropImage.ActivityResult result = CropImage.getActivityResult(data);
-                fileUri = result.getUri();
-            }
-            else {
-                fileUri = imageData;
-            }
-            InputStream imageStream = context.getContentResolver().openInputStream(fileUri);
-            Bitmap selectedImage = BitmapFactory.decodeStream(imageStream);
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-
-            byte[] b;
-            String encImage;
-
-            selectedImage.compress(Bitmap.CompressFormat.JPEG, 100, baos);
-            b = baos.toByteArray();
-            encImage = Base64.encodeToString(b, Base64.NO_WRAP);
-
-            Log.d(UTILS, "camera image size : " + (((encImage.length() / 4) * 3) / 1000));
-
-            if (((encImage.length() / 4) * 3) / 1000 > 400) {
-                int reduceQuality = 10;
-                selectedImage.compress(Bitmap.CompressFormat.JPEG, 100 - reduceQuality, baos);
-                b = baos.toByteArray();
-                encImage = Base64.encodeToString(b, Base64.NO_WRAP);
-                while (((encImage.length() / 4) * 3) / 1000 > 400) {
-                    if (reduceQuality >= 90) {
-                        break;
-                    }
-                    reduceQuality += 10;
-                    baos.reset();
-                    selectedImage.compress(Bitmap.CompressFormat.JPEG, 100 - reduceQuality, baos);
-                    b = baos.toByteArray();
-                    encImage = Base64.encodeToString(b, Base64.NO_WRAP);
-                }
-            }
-
-            Log.d(UTILS, "encoded image size camera : " + (((encImage.length() / 4) * 3) / 1000));
-            {
-                String timeStamp = new SimpleDateFormat("yyyyMMdd_HHmmss", Locale.getDefault()).format(new Date());
-                for (int i = 0; i < callBack.size(); i++) {
-                    if (fileUri != null) {
-                        callBack.get(i).imageUploadCallBack(encImage, "IMG_" + timeStamp + ".jpg", fileUri.getPath());
-                    }
-                }
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-            Bundle params = new Bundle();
-            mFirebaseAnalytics.logEvent("exception_crop_image", params);
-        }
-    }
 
     public static void minimizeApp(Context context) {
         Intent startMain = new Intent(Intent.ACTION_MAIN);
