+++ Frontend/android-native/mobility-app/src/main/java/in/juspay/mobility/app/callbacks/CallBack.java
@@ -7,10 +7,9 @@ public interface CallBack {
 
     void driverCallBack(String notificationType, String notificationData);
 
-    void imageUploadCallBack(String encImage, String filename, String filePath);
-
     void chatCallBack(String message, String sentBy, String time, String len);
 
     void inAppCallBack(String onTapAction);
+
     void bundleUpdatedCallBack(String event, JSONObject description);
 }
