+++ Frontend/android-native/mobility-common/src/main/java/in/juspay/mobility/common/UICallBacks.java
@@ -1,0 +1,5 @@
+package in.juspay.mobility.common;
+
+public interface UICallBacks {
+    void imageUploadCallBack(String encImage, String filename, String filePath);
+}
