+++ Frontend/android-native/mobility-common/src/main/java/in/juspay/mobility/common/services/MobilityAPIResponse.java
@@ -1,4 +1,4 @@
-package in.juspay.mobility.app.services;
+package in.juspay.mobility.common.services;
 
 public class MobilityAPIResponse {
