+++ Frontend/android-native/mobility-common/src/main/java/in/juspay/mobility/common/services/TLSSocketFactory.java
@@ -6,7 +6,7 @@
  * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more details. You should have received a copy of
  * the GNU Affero General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
-package in.juspay.mobility.app;
+package in.juspay.mobility.common.services;
 
 import android.annotation.SuppressLint;
