+++ Frontend/android-native/mobility-common/src/main/res/values-hi/strings.xml
@@ -17,4 +17,5 @@
     <string name="qr_downloaded">QR डाउनलोड!!!</string>
     <string name="upload_image">तस्विर अपलोड करें</string>
     <string name="please_allow_permission_to_capture_the_image">कृपया छवि को कैप्चर करने के लिए अनुमति दें</string>
+    <string name="unable_to_upload_image">छवि अपलोड करने में असमर्थ</string>
 </resources>
