+++ Frontend/android-native/mobility-common/src/main/res/values-ta/strings.xml
@@ -17,4 +17,5 @@
     <string name="qr_downloaded">QR பதிவிறக்கம் !!!</string>
     <string name="upload_image">படத்தை பதிவேற்றம் செய்யவும்</string>
      <string name="please_allow_permission_to_capture_the_image">படத்தைப் பிடிக்க அனுமதிகளை அனுமதிக்கவும்</string>
+    <string name="unable_to_upload_image">படத்தைப் பதிவேற்ற முடியவில்லை</string>
 </resources>
