+++ Frontend/android-native/mobility-common/src/main/res/values/attrs.xml
