+++ Frontend/android-native/mobility-common/src/main/res/values/strings.xml
@@ -18,4 +18,5 @@
     <string name="upload_image">Upload Image</string>
     <string name="TIME_STAMP_FILE_UPLOAD" translatable="false">TIME_STAMP_FILE_UPLOAD</string>
     <string name="please_allow_permission_to_capture_the_image">Please allow permissions to capture the image</string>
+    <string name="unable_to_upload_image">Unable to Upload Image</string>
 </resources>
