+++ Frontend/android-native/mobility-customer/build.gradle
@@ -46,7 +46,6 @@ dependencies {
 
     implementation project(':mobility-common')
     api 'in.juspay:hyper:2.1.20-nammayatri.13'
-    compileOnly project(':mobility-app')
     compileOnly 'com.google.maps.android:android-maps-utils:2.4.0'
     compileOnly 'com.google.android.gms:play-services-maps:18.1.0'
     compileOnly 'com.google.android.gms:play-services-location:21.0.1'
